Refactoring properties to use descriptors properly This will allow to use multiple instances of device
